requirements.txt
asgiref==3.4.1
boto3==1.18.33
botocore==1.21.33
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.6
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-storages==1.11.1
django-taggit==1.5.1
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
Pillow==8.3.1
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
